drastic restructuring, plus new AI for enemies as an experiment towards our modular AI system
eonz library was expanded, most significant is eonz.namespace which changes the way we load modules from the traditional `require`. Modules are now lazily loaded as they are referenced by name as indices of a "namespace" object for their folder.

TODO: documentation for a lot of this
TODO: split the strategies into tactics and strategies, use messaging system/callbacks to control switching between them