reworked strategy system into modular tactics with callbacks
This allows strategies to be described at a higher level asynchronously.

EnemyController is now TacticsController. TacticsController passes around an "environment" table to the tactics set on it that is persistent over the life of the controller. It can be used to store state, though it probably should not be. We should prefer to store strategic state in a closure that defines the high level strategy. (see ZombieStrat.lua)